- added priest abilities - finished ritual code - removed lore inconsistencies - buffed and reworked missile weapons - changed lore of the warrior's seal, - arena changes, adding interest points

+++ pixel-towers-master/core/src/main/java/com/fixakathefix/towerpixeldungeon/actors/mobs/ArmoredStatue.java
@@ -158,7 +158,7 @@ public class ArmoredStatue extends Statue {
 
 	@Override
 	public void die(Object cause) {
-		Dungeon.level.drop( armor, pos ).sprite.drop();
+		if (willdropweapon) Dungeon.level.drop( armor, pos ).sprite.drop();
 		super.die(cause);
 	}
